Move mqtt queue models to mqtt queue file
