+++ src/isar/services/service_connections/mqtt/robot_heartbeat_publisher.py
@@ -2,8 +2,7 @@ import time
 from datetime import UTC, datetime
 
 from isar.config.settings import settings
-from isar.models.mqtt_queue import MQTTQueue
-from robot_interface.telemetry.mqtt_client import props_expiry
+from isar.models.mqtt_queue import MQTTQueue, props_expiry
 from robot_interface.telemetry.payloads import RobotHeartbeatPayload
