+++ tests/test_mocks/robot_interface.py
@@ -5,6 +5,7 @@ from queue import Queue
 from alitra import Frame, Orientation, Pose, Position
 
 from isar.models.events import Event
+from isar.models.mqtt_queue import MQTTQueueMessage
 from isar.state_machine.states_enum import States
 from robot_interface.models.exceptions.robot_exceptions import (
     RobotCommunicationException,
@@ -16,10 +17,7 @@ from robot_interface.models.mission.status import MissionStatus, RobotStatus, Ta
 from robot_interface.models.mission.task import InspectionTask
 from robot_interface.models.robots.media import MediaConfig, MediaConnectionType
 from robot_interface.robot_interface import RobotInterface
-from robot_interface.telemetry.mqtt_client import (
-    MQTTQueueMessage,
-    MqttTelemetryPublisher,
-)
+from robot_interface.telemetry.mqtt_client import MqttTelemetryPublisher
 from tests.test_mocks.inspection import stub_image_metadata
 
 _ROBOT_FRAME = Frame(name="robot")
